fix(theme)!: ship ExtendedTheme in published types, break circular augmentation
useExtendedTheme() previously resolved to `any` for consumers because the
published types imported ExtendedTheme from an unshipped file. This ships the
type, which is a breaking change for typed consumers.

BREAKING CHANGE: useExtendedTheme() now returns a typed ExtendedTheme instead of
`any`, and the package's MUI augmentation (Palette.tertiary) now reaches all
consumers for the first time. Code relying on the previous `any` theme, or
constructing a Palette literal, may need type changes. Runtime is unchanged.
